move unused func to test file
