+++ smartcontract/service/native/ontid/group.go
@@ -22,7 +22,6 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/ontio/ontology/account"
 	"github.com/ontio/ontology/common"
 	"github.com/ontio/ontology/smartcontract/service/native"
 	"github.com/ontio/ontology/smartcontract/service/native/utils"
@@ -39,26 +38,6 @@ type Group struct {
 func (g *Group) ToJson() []byte {
 	j, _ := json.Marshal(g)
 	return j
-}
-
-func (g *Group) Serialize() []byte {
-	sink := common.NewZeroCopySink(nil)
-	utils.EncodeVarUint(sink, uint64(len(g.Members)))
-	for _, m := range g.Members {
-		switch t := m.(type) {
-		case []byte:
-			if !account.VerifyID(string(t)) {
-				panic("invalid ont id format")
-			}
-			sink.WriteVarBytes(t)
-		case *Group:
-			sink.WriteVarBytes(t.Serialize())
-		default:
-			panic("invalid member type")
-		}
-	}
-	utils.EncodeVarUint(sink, uint64(g.Threshold))
-	return sink.Bytes()
 }
 
 func rDeserialize(data []byte, depth uint) (*Group, error) {
